fix: move PublishedDeclaration type and extractor to server-safe file
The Payload server context (ActionPlans.ts → publish-comparison.ts)
was importing from PublishedDeclarationTemplate.tsx which transitively
pulls in DeclarationMarkdownToJsx → tss.withName, a browser-only call,
causing a runtime crash in middleware/server.

Extract PublishedDeclaration type and extractDeclarationContentToPublish
into src/utils/declaration-content.ts (no React imports). Re-export both
from PublishedDeclarationTemplate.tsx for backward compatibility with
page files (preview.tsx, publish.tsx).
