+++ src/components/declaration/PublishedDeclarationTemplate.tsx
@@ -1,86 +1,8 @@
 import DeclarationMarkdownToJsx from "~/components/declaration/DeclarationMarkdownToJsx";
-import {
-	appKindOptions,
-	rgaaVersionOptions,
-	testEnvironmentOptions,
-	toolOptions,
-} from "~/payload/selectOptions";
-import type { PopulatedDeclaration } from "~/server/api/utils/payload-helper";
 import { getConformityStatus } from "~/utils/declaration-helper";
-
-export const extractDeclarationContentToPublish = (
-	declaration: PopulatedDeclaration,
-): PublishedDeclaration => {
-	return {
-		name: declaration.name ?? "",
-		entityName: declaration.entity?.name ?? "",
-		actionPlan: {
-			currentYearSchemaUrl: declaration?.actionPlan?.currentYearSchemaUrl ?? "",
-			previousYearsSchemaUrl:
-				declaration?.actionPlan?.previousYearsSchemaUrl ?? "",
-		},
-		appKindLabel:
-			appKindOptions.find((kind) => kind.value === declaration.app_kind)
-				?.label ?? "",
-		url: declaration?.url ?? "",
-		audit: {
-			rgaa_version:
-				rgaaVersionOptions.find(
-					(option) => option.value === declaration?.audit?.rgaa_version,
-				)?.label ?? "RGAA 4",
-			realised_by: declaration.audit?.realisedBy ?? "",
-			rate: declaration.audit?.rate ?? 0,
-			nonCompliantElements: declaration?.audit?.nonCompliantElements ?? "",
-			disproportionnedCharge: declaration?.audit?.disproportionnedCharge ?? "",
-			optionalElements: declaration?.audit?.optionalElements ?? "",
-			compliantElements: declaration?.audit?.compliantElements ?? "",
-			technologies: declaration?.audit?.technologies ?? [],
-			testEnvironments:
-				(declaration?.audit?.testEnvironments ?? [])?.map(
-					(env) =>
-						testEnvironmentOptions.find((option) => option.value === env.name)
-							?.label ?? env.name,
-				) ?? [],
-			usedTools:
-				(declaration?.audit?.usedTools ?? [])?.map(
-					(tool) =>
-						toolOptions.find((option) => option.value === tool.name)?.label ??
-						tool.name,
-				) ?? [],
-		},
-		contact: {
-			url: declaration.contact?.url ?? "",
-			email: declaration.contact?.email ?? "",
-		},
-	};
-};
-
-export type PublishedDeclaration = {
-	name: string;
-	entityName: string;
-	actionPlan: {
-		currentYearSchemaUrl: string;
-		previousYearsSchemaUrl: string;
-	};
-	appKindLabel: string;
-	url: string;
-	audit: {
-		rgaa_version: string | undefined;
-		realised_by: string;
-		rate: number;
-		nonCompliantElements: string | null;
-		disproportionnedCharge: string | null;
-		optionalElements: string | null;
-		compliantElements: string;
-		technologies: { name: string }[];
-		testEnvironments: string[];
-		usedTools: string[];
-	};
-	contact: {
-		url: string | null;
-		email: string | null;
-	};
-};
+export type { PublishedDeclaration } from "~/utils/declaration-content";
+export { extractDeclarationContentToPublish } from "~/utils/declaration-content";
+import type { PublishedDeclaration } from "~/utils/declaration-content";
 
 export default function PublishedDeclarationTemplate({
 	declaration,
